Corrige FK de actas_reunion: evento_id BIGINT para coincidir con eventos.id
En una BD nueva la migracion fallaba con errno 150 (Foreign key constraint
is incorrectly formed) porque evento_id era INT y eventos.id es BIGINT.

+++ backend/config/migrations.js
@@ -137,7 +137,7 @@ async function asegurarTablaActasReunion() {
   await ejecutar(`
     CREATE TABLE IF NOT EXISTS actas_reunion (
       id            INT AUTO_INCREMENT PRIMARY KEY,
-      evento_id     INT NOT NULL,
+      evento_id     BIGINT NOT NULL,
       titulo        VARCHAR(200) NOT NULL,
       contenido     TEXT NOT NULL,
       archivo_path  VARCHAR(255) NULL,
